% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{skeleton_update}
\alias{skeleton_update}
\title{Copying vignette Rmd as R Markdown document template}
\usage{
skeleton_update(rmd)
}
\arguments{
\item{rmd}{vignette *.Rmd path. Usually: "vignettes/<filename>.Rmd"}
}
\value{
Path to the newly created/update Rmarkdown template.
}
\description{
Copying vignette Rmd as R Markdown document template
}
\examples{
sprthis(wfName = "SPRtest", analysis = "SPRtest", path = tempdir())
path <- file.path(tempdir(), "vignette/SPRtest.Rmd")
skeleton_update(path)
}
