% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprthis.R
\name{sprthis}
\alias{sprthis}
\title{Create SPR Workflow package template}
\usage{
sprthis(wfName = "SPRtest", analysis, path = tempdir(), baseurl = "<baseurl>")
}
\arguments{
\item{wfName}{character vector with the new SPR Workflow package name.}

\item{analysis}{character vector withe the analysis name.}

\item{path}{A path. If it exists, it is used. If it does not exist, it will be used the temporary path..}

\item{baseurl}{baseurl for the package}
}
\value{
Path to the newly created package.
}
\description{
Create SPR Workflow package template
}
\examples{
sprthis(wfName = "SPRtest", analysis = "SPRtest", path = tempdir())
}
